+++ l2geth/core/state_transition.go
@@ -77,8 +77,7 @@ type Message interface {
 	Data() []byte
 	L1MessageSender() *common.Address
 	L1BlockNumber() *big.Int
-	QueueOrigin() *big.Int
-	SignatureHashType() types.SignatureHashType
+	QueueOrigin() types.QueueOrigin
 }
 
 // IntrinsicGas computes the 'intrinsic gas' for a message with the given data.
@@ -185,9 +184,7 @@ func (st *StateTransition) preCheck() error {
 		if nonce < st.msg.Nonce() {
 			if vm.UsingOVM {
 				// The nonce never increments for L1ToL2 txs
-				qo := st.msg.QueueOrigin()
-				l1ToL2 := uint64(types.QueueOriginL1ToL2)
-				if qo != nil && qo.Uint64() == l1ToL2 {
+				if st.msg.QueueOrigin() == types.QueueOriginL1ToL2 {
 					return st.buyGas()
 				}
 			}
@@ -252,7 +249,7 @@ func (st *StateTransition) TransitionDb() (ret []byte, usedGas uint64, failed bo
 			l1MessageSender = msg.L1MessageSender().Hex()
 		}
 		if st.evm.EthCallSender == nil {
-			log.Debug("Applying transaction", "ID", st.evm.Id, "from", sender.Address().Hex(), "to", to, "nonce", msg.Nonce(), "gasPrice", msg.GasPrice().Uint64(), "gasLimit", msg.Gas(), "l1MessageSender", l1MessageSender, "data", hexutil.Encode(msg.Data()))
+			log.Debug("Applying transaction", "ID", st.evm.Id, "from", sender.Address().Hex(), "to", to, "nonce", msg.Nonce(), "gasPrice", msg.GasPrice().Uint64(), "gasLimit", msg.Gas(), "value", msg.Value().Uint64(), "l1MessageSender", l1MessageSender, "data", hexutil.Encode(msg.Data()))
 		}
 	}
 
@@ -283,6 +280,9 @@ func (st *StateTransition) TransitionDb() (ret []byte, usedGas uint64, failed bo
 }
 
 func (st *StateTransition) refundGas() {
+	if vm.UsingOVM {
+		return
+	}
 	// Apply refund counter, capped to half of the used gas.
 	refund := st.gasUsed() / 2
 	if refund > st.state.GetRefund() {