+++ packages/contracts/bin/deploy.ts
@@ -1,5 +1,3 @@
-#!/usr/bin/env ts-node-script
-
 import { Wallet } from 'ethers'
 import path from 'path'
 import dirtree from 'directory-tree'
@@ -18,10 +16,42 @@ process.env.CONTRACTS_RPC_URL =
 import hre from 'hardhat'
 
 const sequencer = new Wallet(process.env.SEQUENCER_PRIVATE_KEY)
+const proposer = new Wallet(process.env.PROPOSER_PRIVATE_KEY)
 const deployer = new Wallet(process.env.DEPLOYER_PRIVATE_KEY)
+const relayer = new Wallet(process.env.RELAYER_PRIVATE_KEY)
+
+const parseEnv = () => {
+  const ensure = (env, type) => {
+    if (typeof process.env[env] === 'undefined') {
+      return undefined
+    }
+    if (type === 'number') {
+      return parseInt(process.env[env], 10)
+    }
+    return process.env[env]
+  }
+
+  return {
+    l1BlockTimeSeconds: ensure('BLOCK_TIME_SECONDS', 'number'),
+    ctcForceInclusionPeriodSeconds: ensure(
+      'FORCE_INCLUSION_PERIOD_SECONDS',
+      'number'
+    ),
+    ctcMaxTransactionGasLimit: ensure('MAX_TRANSACTION_GAS_LIMIT', 'number'),
+    emMinTransactionGasLimit: ensure('MIN_TRANSACTION_GAS_LIMIT', 'number'),
+    emMaxtransactionGasLimit: ensure('MAX_TRANSACTION_GAS_LIMIT', 'number'),
+    emMaxGasPerQueuePerEpoch: ensure('MAX_GAS_PER_QUEUE_PER_EPOCH', 'number'),
+    emSecondsPerEpoch: ensure('ECONDS_PER_EPOCH', 'number'),
+    emOvmChainId: ensure('CHAIN_ID', 'number'),
+    sccFraudProofWindow: ensure('FRAUD_PROOF_WINDOW_SECONDS', 'number'),
+    sccSequencerPublishWindow: ensure(
+      'SEQUENCER_PUBLISH_WINDOW_SECONDS',
+      'number'
+    ),
+  }
+}
 
 const main = async () => {
-
   const config = parseEnv()
 
   await hre.run('deploy', {
@@ -36,8 +66,8 @@ const main = async () => {
     sccFraudProofWindow: config.sccFraudProofWindow,
     sccSequencerPublishWindow: config.sccFraudProofWindow,
     ovmSequencerAddress: sequencer.address,
-    ovmProposerAddress: sequencer.address,
-    ovmRelayerAddress: sequencer.address,
+    ovmProposerAddress: proposer.address,
+    ovmRelayerAddress: relayer.address,
     ovmAddressManagerOwner: deployer.address,
     noCompile: process.env.NO_COMPILE ? true : false,
   })
@@ -46,26 +76,33 @@ const main = async () => {
   // update our CI so this is no longer necessary. But I'm adding it for backwards compat so we can
   // get the hardhat-deploy stuff merged. Woot.
   const nicknames = {
-    'Lib_AddressManager': 'AddressManager',
-    'mockOVM_BondManager': 'OVM_BondManager'
+    Lib_AddressManager: 'AddressManager',
+    mockOVM_BondManager: 'OVM_BondManager',
   }
 
   const contracts: any = dirtree(
     path.resolve(__dirname, `../deployments/custom`)
-  ).children.filter((child) => {
-    return child.extension === '.json'
-  }).reduce((contracts, child) => {
-    const contractName = child.name.replace('.json', '')
-    const artifact = require(path.resolve(__dirname, `../deployments/custom/${child.name}`))
-    contracts[nicknames[contractName] || contractName] = artifact.address
-    return contracts
-  }, {})
+  )
+    .children.filter((child) => {
+      return child.extension === '.json'
+    })
+    .reduce((contractsAccumulator, child) => {
+      const contractName = child.name.replace('.json', '')
+      // eslint-disable-next-line @typescript-eslint/no-var-requires
+      const artifact = require(path.resolve(
+        __dirname,
+        `../deployments/custom/${child.name}`
+      ))
+      contractsAccumulator[nicknames[contractName] || contractName] =
+        artifact.address
+      return contractsAccumulator
+    }, {})
 
   contracts.OVM_Sequencer = await sequencer.getAddress()
   contracts.Deployer = await deployer.getAddress()
 
   const addresses = JSON.stringify(contracts, null, 2)
-  const dumpsPath = path.resolve(__dirname, "../dist/dumps")
+  const dumpsPath = path.resolve(__dirname, '../dist/dumps')
   if (!fs.existsSync(dumpsPath)) {
     fs.mkdirSync(dumpsPath)
   }
@@ -81,26 +118,3 @@ main()
     )
     process.exit(1)
   })
-
-function parseEnv() {
-  function ensure(env, type) {
-    if (typeof process.env[env] === 'undefined')
-      return undefined
-    if (type === 'number')
-      return parseInt(process.env[env], 10)
-    return process.env[env]
-  }
-
-  return {
-    l1BlockTimeSeconds: ensure('BLOCK_TIME_SECONDS', 'number'),
-    ctcForceInclusionPeriodSeconds: ensure('FORCE_INCLUSION_PERIOD_SECONDS', 'number'),
-    ctcMaxTransactionGasLimit: ensure('MAX_TRANSACTION_GAS_LIMIT', 'number'),
-    emMinTransactionGasLimit: ensure('MIN_TRANSACTION_GAS_LIMIT', 'number'),
-    emMaxtransactionGasLimit: ensure('MAX_TRANSACTION_GAS_LIMIT', 'number'),
-    emMaxGasPerQueuePerEpoch: ensure('MAX_GAS_PER_QUEUE_PER_EPOCH', 'number'),
-    emSecondsPerEpoch: ensure('ECONDS_PER_EPOCH', 'number'),
-    emOvmChainId: ensure('CHAIN_ID', 'number'),
-    sccFraudProofWindow: ensure('FRAUD_PROOF_WINDOW_SECONDS', 'number'),
-    sccSequencerPublishWindow: ensure('SEQUENCER_PUBLISH_WINDOW_SECONDS', 'number'),
-  }
-}