+++ packages/data-transport-layer/src/types/database-types.ts
@@ -5,19 +5,18 @@ export interface DecodedSequencerBatchTransaction {
     v: number
   }
   value: string
-  gasLimit: number
-  gasPrice: number
-  nonce: number
+  gasLimit: string
+  gasPrice: string
+  nonce: string
   target: string
   data: string
-  type: number
 }
 
 export interface EnqueueEntry {
   index: number
   target: string
   data: string
-  gasLimit: number
+  gasLimit: string
   origin: string
   blockNumber: number
   timestamp: number
@@ -29,13 +28,12 @@ export interface TransactionEntry {
   data: string
   blockNumber: number
   timestamp: number
-  gasLimit: number
+  gasLimit: string
   target: string
   origin: string
   value: string
   queueOrigin: 'sequencer' | 'l1'
   queueIndex: number | null
-  type: 'EIP155' | 'ETH_SIGN' | null
   decoded: DecodedSequencerBatchTransaction | null
   confirmed: boolean
 }