+++ packages/message-relayer/src/service.ts
@@ -2,11 +2,11 @@
 import { Contract, ethers, Wallet, BigNumber, providers } from 'ethers'
 import * as rlp from 'rlp'
 import { MerkleTree } from 'merkletreejs'
+import fetch from 'node-fetch';
 
 /* Imports: Internal */
 import { fromHexString, sleep } from '@eth-optimism/core-utils'
-import { BaseService } from '@eth-optimism/common-ts'
-import SpreadSheet from './spreadsheet'
+import { Logger, BaseService, Metrics } from '@eth-optimism/common-ts'
 
 import { loadContract, loadContractFromManager } from '@eth-optimism/contracts'
 import { StateRootBatchHeader, SentMessage, SentMessageProof } from './types'
@@ -42,9 +42,16 @@ interface MessageRelayerOptions {
   // Number of blocks within each getLogs query - max is 2000
   getLogsInterval?: number
 
-  // Append txs to a spreadsheet instead of submitting transactions
-  spreadsheetMode?: boolean
-  spreadsheet?: SpreadSheet
+  // A custom logger to transport logs via; default STDOUT
+  logger?: Logger
+
+  // A custom metrics tracker to manage metrics; default undefined
+  metrics?: Metrics
+
+  // filter
+  filterEndpoint?: string
+
+  filterPollingInterval?: number
 }
 
 const optionSettings = {
@@ -54,16 +61,13 @@ const optionSettings = {
   l2BlockOffset: { default: 1 },
   l1StartOffset: { default: 0 },
   getLogsInterval: { default: 2000 },
-  spreadsheetMode: { default: false },
+  filterPollingInterval: { default: 60000 }
 }
 
 export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
   constructor(options: MessageRelayerOptions) {
     super('Message_Relayer', options, optionSettings)
   }
-
-  protected spreadsheetMode: boolean
-  protected spreadsheet: SpreadSheet
 
   private state: {
     lastFinalizedTxHeight: number
@@ -75,10 +79,19 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
     OVM_L1CrossDomainMessenger: Contract
     OVM_L2CrossDomainMessenger: Contract
     OVM_L2ToL1MessagePasser: Contract
+    filter: Array<any>
+    lastFilterPollingTimestamp: number
   }
 
   protected async _init(): Promise<void> {
-    this.logger.info('Initializing message relayer', { options: this.options })
+    this.logger.info('Initializing message relayer', {
+      relayGasLimit: this.options.relayGasLimit,
+      fromL2TransactionIndex: this.options.fromL2TransactionIndex,
+      pollingInterval: this.options.pollingInterval,
+      l2BlockOffset: this.options.l2BlockOffset,
+      getLogsInterval: this.options.getLogsInterval,
+      filterPollingInterval: this.options.filterPollingInterval,
+    })
     // Need to improve this, sorry.
     this.state = {} as any
 
@@ -134,21 +147,19 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
 
     this.logger.info('Connected to all contracts.')
 
-    if (this.options.spreadsheetMode) {
-      this.logger.info('Running in spreadsheet mode')
-    }
-
     this.state.lastQueriedL1Block = this.options.l1StartOffset
     this.state.eventCache = []
 
     this.state.lastFinalizedTxHeight = this.options.fromL2TransactionIndex || 0
     this.state.nextUnfinalizedTxHeight =
       this.options.fromL2TransactionIndex || 0
+    this.state.lastFilterPollingTimestamp = 0
   }
 
   protected async _start(): Promise<void> {
     while (this.running) {
       await sleep(this.options.pollingInterval)
+      await this._getFilter()
 
       try {
         // Check that the correct address is set in the address manager
@@ -190,6 +201,17 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
             { batchSize: size }
           )
           this.state.nextUnfinalizedTxHeight += size
+
+          // Only deal with ~1000 transactions at a time so we can limit the amount of stuff we
+          // need to keep in memory. We operate on full batches at a time so the actual amount
+          // depends on the size of the batches we're processing.
+          const numTransactionsToProcess =
+            this.state.nextUnfinalizedTxHeight -
+            this.state.lastFinalizedTxHeight
+
+          if (numTransactionsToProcess > 1000) {
+            break
+          }
         }
 
         this.logger.info('Found finalized transactions', {
@@ -203,18 +225,17 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
           this.state.nextUnfinalizedTxHeight
         )
 
-        if (messages.length === 0) {
-          this.logger.info('Did not find any L2->L1 messages', {
-            retryAgainInS: Math.floor(this.options.pollingInterval / 1000),
-          })
-        }
-
         for (const message of messages) {
           this.logger.info('Found a message sent during transaction', {
             index: message.parentTransactionIndex,
           })
           if (await this._wasMessageRelayed(message)) {
             this.logger.info('Message has already been relayed, skipping.')
+            continue
+          }
+
+          if (this.state.filter.includes(message.target)) {
+            this.logger.info('Message not intended for target, skipping.')
             continue
           }
 
@@ -229,6 +250,27 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
           await this._relayMessageToL1(message, proof)
         }
 
+        if (messages.length === 0) {
+          this.logger.info('Did not find any L2->L1 messages', {
+            retryAgainInS: Math.floor(this.options.pollingInterval / 1000),
+          })
+        } else {
+          // Clear the event cache to avoid keeping every single event in memory and eventually
+          // getting OOM killed. Messages are already sorted in ascending order so the last message
+          // will have the highest batch index.
+          const lastMessage = messages[messages.length - 1]
+
+          // Find the batch corresponding to the last processed message.
+          const lastProcessedBatch = await this._getStateBatchHeader(
+            lastMessage.parentTransactionIndex
+          )
+
+          // Remove any events from the cache for batches that should've been processed by now.
+          this.state.eventCache = this.state.eventCache.filter((event) => {
+            return event.args._batchIndex > lastProcessedBatch.batch.batchIndex
+          })
+        }
+
         this.logger.info(
           'Finished searching through newly finalized transactions',
           {
@@ -236,21 +278,36 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
           }
         )
       } catch (err) {
-        this.logger.error('Caught an unhandled error', { err })
+        this.logger.error('Caught an unhandled error', {
+          message: err.toString(),
+          stack: err.stack,
+          code: err.code,
+        })
       }
     }
   }
 
-  private async _getStateBatchHeader(
-    height: number
-  ): Promise<
+  private async _getStateBatchHeader(height: number): Promise<
     | {
         batch: StateRootBatchHeader
         stateRoots: string[]
       }
     | undefined
   > {
-    const filter = this.state.OVM_StateCommitmentChain.filters.StateBatchAppended()
+    const getStateBatchAppendedEventForIndex = (
+      txIndex: number
+    ): ethers.Event => {
+      return this.state.eventCache.find((cachedEvent) => {
+        const prevTotalElements = cachedEvent.args._prevTotalElements.toNumber()
+        const batchSize = cachedEvent.args._batchSize.toNumber()
+
+        // Height should be within the bounds of the batch.
+        return (
+          txIndex >= prevTotalElements &&
+          txIndex < prevTotalElements + batchSize
+        )
+      })
+    }
 
     let startingBlock = this.state.lastQueriedL1Block
     while (
@@ -262,50 +319,48 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
         endBlock: startingBlock + this.options.getLogsInterval,
       })
 
-      const events: ethers.Event[] = await this.state.OVM_StateCommitmentChain.queryFilter(
-        filter,
-        startingBlock,
-        startingBlock + this.options.getLogsInterval
-      )
+      const events: ethers.Event[] =
+        await this.state.OVM_StateCommitmentChain.queryFilter(
+          this.state.OVM_StateCommitmentChain.filters.StateBatchAppended(),
+          startingBlock,
+          startingBlock + this.options.getLogsInterval
+        )
 
       this.state.eventCache = this.state.eventCache.concat(events)
       startingBlock += this.options.getLogsInterval
+
+      // We need to stop syncing early once we find the event we're looking for to avoid putting
+      // *all* events into memory at the same time. Otherwise we'll get OOM killed.
+      if (getStateBatchAppendedEventForIndex(height) !== undefined) {
+        break
+      }
     }
 
-    // tslint:disable-next-line
-    const event = this.state.eventCache.find((event) => {
-      return (
-        event.args._prevTotalElements.toNumber() <= height &&
-        event.args._prevTotalElements.toNumber() +
-          event.args._batchSize.toNumber() >
-          height
-      )
-    })
+    const event = getStateBatchAppendedEventForIndex(height)
+    if (event === undefined) {
+      return undefined
+    }
 
-    if (event) {
-      const transaction = await this.options.l1RpcProvider.getTransaction(
-        event.transactionHash
-      )
-      const [
-        stateRoots,
-      ] = this.state.OVM_StateCommitmentChain.interface.decodeFunctionData(
+    const transaction = await this.options.l1RpcProvider.getTransaction(
+      event.transactionHash
+    )
+
+    const [stateRoots] =
+      this.state.OVM_StateCommitmentChain.interface.decodeFunctionData(
         'appendStateBatch',
         transaction.data
       )
 
-      return {
-        batch: {
-          batchIndex: event.args._batchIndex,
-          batchRoot: event.args._batchRoot,
-          batchSize: event.args._batchSize,
-          prevTotalElements: event.args._prevTotalElements,
-          extraData: event.args._extraData,
-        },
-        stateRoots,
-      }
+    return {
+      batch: {
+        batchIndex: event.args._batchIndex,
+        batchRoot: event.args._batchRoot,
+        batchSize: event.args._batchSize,
+        prevTotalElements: event.args._prevTotalElements,
+        extraData: event.args._extraData,
+      },
+      stateRoots,
     }
-
-    return
   }
 
   private async _isTransactionFinalized(height: number): Promise<boolean> {
@@ -324,6 +379,15 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
     ))
   }
 
+  /**
+   * Returns all sent message events between some start height (inclusive) and an end height
+   * (exclusive).
+   *
+   * @param startHeight Start height to start finding messages from.
+   * @param endHeight End height to finish finding messages at.
+   * @returns All sent messages between start and end height, sorted by transaction index in
+   * ascending order.
+   */
   private async _getSentMessages(
     startHeight: number,
     endHeight: number
@@ -335,12 +399,13 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
       endHeight + this.options.l2BlockOffset - 1
     )
 
-    return events.map((event) => {
+    const messages = events.map((event) => {
       const message = event.args.message
-      const decoded = this.state.OVM_L2CrossDomainMessenger.interface.decodeFunctionData(
-        'relayMessage',
-        message
-      )
+      const decoded =
+        this.state.OVM_L2CrossDomainMessenger.interface.decodeFunctionData(
+          'relayMessage',
+          message
+        )
 
       return {
         target: decoded._target,
@@ -352,6 +417,11 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
         parentTransactionIndex: event.blockNumber - this.options.l2BlockOffset,
         parentTransactionHash: event.transactionHash,
       }
+    })
+
+    // Sort in ascending order based on tx index and return.
+    return messages.sort((a, b) => {
+      return a.parentTransactionIndex - b.parentTransactionIndex
     })
   }
 
@@ -433,63 +503,12 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
     message: SentMessage,
     proof: SentMessageProof
   ): Promise<void> {
-    if (this.options.spreadsheetMode) {
-      try {
-        await this.options.spreadsheet.addRow({
-          target: message.target,
-          sender: message.sender,
-          message: message.message,
-          messageNonce: message.messageNonce.toString(),
-          encodedMessage: message.encodedMessage,
-          encodedMessageHash: message.encodedMessageHash,
-          parentTransactionIndex: message.parentTransactionIndex,
-          parentTransactionHash: message.parentTransactionIndex,
-          stateRoot: proof.stateRoot,
-          batchIndex: proof.stateRootBatchHeader.batchIndex.toString(),
-          batchRoot: proof.stateRootBatchHeader.batchRoot,
-          batchSize: proof.stateRootBatchHeader.batchSize.toString(),
-          prevTotalElements: proof.stateRootBatchHeader.prevTotalElements.toString(),
-          extraData: proof.stateRootBatchHeader.extraData,
-          index: proof.stateRootProof.index,
-          siblings: proof.stateRootProof.siblings.join(','),
-          stateTrieWitness: proof.stateTrieWitness.toString('hex'),
-          storageTrieWitness: proof.storageTrieWitness.toString('hex'),
-        })
-        this.logger.info('Submitted relay message to spreadsheet')
-      } catch (e) {
-        this.logger.error('Cannot submit message to spreadsheet')
-        this.logger.error(e.message)
-      }
-    } else {
-      try {
-        this.logger.info(
-          'Dry-run, checking to make sure proof would succeed...'
-        )
+    try {
+      this.logger.info('Dry-run, checking to make sure proof would succeed...')
 
-        await this.state.OVM_L1CrossDomainMessenger.connect(
-          this.options.l1Wallet
-        ).callStatic.relayMessage(
-          message.target,
-          message.sender,
-          message.message,
-          message.messageNonce,
-          proof,
-          {
-            gasLimit: this.options.relayGasLimit,
-          }
-        )
-
-        this.logger.info(
-          'Proof should succeed. Submitting for real this time...'
-        )
-      } catch (err) {
-        this.logger.error('Proof would fail, skipping', { err })
-        return
-      }
-
-      const result = await this.state.OVM_L1CrossDomainMessenger.connect(
+      await this.state.OVM_L1CrossDomainMessenger.connect(
         this.options.l1Wallet
-      ).relayMessage(
+      ).callStatic.relayMessage(
         message.target,
         message.sender,
         message.message,
@@ -500,25 +519,78 @@ export class MessageRelayerService extends BaseService<MessageRelayerOptions> {
         }
       )
 
-      this.logger.info('Relay message transaction sent', {
-        transactionHash: result,
+      this.logger.info('Proof should succeed. Submitting for real this time...')
+    } catch (err) {
+      this.logger.error('Proof would fail, skipping', {
+        message: err.toString(),
+        stack: err.stack,
+        code: err.code,
       })
+      return
+    }
 
-      try {
-        const receipt = await result.wait()
-
-        this.logger.info('Relay message included in block', {
-          transactionHash: receipt.transactionHash,
-          blockNumber: receipt.blockNumber,
-          gasUsed: receipt.gasUsed.toString(),
-          confirmations: receipt.confirmations,
-          status: receipt.status,
-        })
-      } catch (err) {
-        this.logger.error('Real relay attempt failed, skipping.', { err })
-        return
+    const result = await this.state.OVM_L1CrossDomainMessenger.connect(
+      this.options.l1Wallet
+    ).relayMessage(
+      message.target,
+      message.sender,
+      message.message,
+      message.messageNonce,
+      proof,
+      {
+        gasLimit: this.options.relayGasLimit,
       }
-      this.logger.info('Message successfully relayed to Layer 1!')
+    )
+
+    this.logger.info('Relay message transaction sent', {
+      transactionHash: result,
+    })
+
+    try {
+      const receipt = await result.wait()
+
+      this.logger.info('Relay message included in block', {
+        transactionHash: receipt.transactionHash,
+        blockNumber: receipt.blockNumber,
+        gasUsed: receipt.gasUsed.toString(),
+        confirmations: receipt.confirmations,
+        status: receipt.status,
+      })
+    } catch (err) {
+      this.logger.error('Real relay attempt failed, skipping.', {
+        message: err.toString(),
+        stack: err.stack,
+        code: err.code,
+      })
+      return
+    }
+    this.logger.info('Message successfully relayed to Layer 1!')
+  }
+
+  private async _getFilter(): Promise<void> {
+    try {
+      if (this.options.filterEndpoint) {
+        if (this.state.lastFilterPollingTimestamp === 0 ||
+          new Date().getTime() > this.state.lastFilterPollingTimestamp + this.options.filterPollingInterval
+        ) {
+          const response = await fetch(this.options.filterEndpoint)
+          const filter = await response.json()
+
+          // export L1LIQPOOL=$(echo $ADDRESSES | jq -r '.L1LiquidityPool')
+          // export L1M=$(echo $ADDRESSES | jq -r '.L1Message')
+          // echo '["'$L1LIQPOOL'", "'$L1M'"]' > dist/dumps/whitelist.json
+          const filterSelect = [ filter.L1LiquidityPool, filter.L1Message ]
+
+          this.state.lastFilterPollingTimestamp = new Date().getTime()
+          this.state.filter = filterSelect
+          this.logger.info('Found the filter', { filterSelect })
+        }
+      } else {
+        this.state.filter = [];
+      }
+    } catch {
+      this.logger.info('Failed to fetch the filter')
+      this.state.filter = [];
     }
   }
 }